feat: implement pagination working correctly while fetching data
