+++ async-race/src/components/garage/garage.ts
@@ -1,8 +1,6 @@
-import { Numbers } from '../../enums/numbers';
 import { Urls } from '../../enums/urls';
 import type { INewCar } from '../../types/types';
 import { BaseComponent } from '../../utils/base-component';
-import { Pagination } from '../pagination/pagination';
 import { Controls } from './controls-component/controls';
 import { Track } from './track-component/track';
 import type { IGarage } from './types/garage-types';
@@ -12,20 +10,17 @@ export class Garage extends BaseComponent implements IGarage {
   constructor(
     public controls = new Controls(),
     public track = new Track(),
-    public pagination = new Pagination(),
   ) {
     super(garageView.garage);
 
     this.getElement().append(
       this.controls.getElement(),
       this.track.getElement(),
-      this.pagination.getElement(),
     );
 
     this.addCreateCarHandler();
     this.addEnableUpgradeCarHandler();
     this.addUpgradeCarHandler();
-    this.addPaginationHandler();
   }
 
   private addCreateCarHandler(): void {
@@ -99,26 +94,5 @@ export class Garage extends BaseComponent implements IGarage {
         });
     };
     this.controls.upgradeCarBtn.addEventListener('click', upgradeCarHandler);
-  }
-
-  private addPaginationHandler(): void {
-    const paginationNextHandler = (): void => {
-      if (
-        this.pagination.currentPage < Math.ceil(this.track.carsList.length / Numbers.CARS_ON_PAGE)
-      ) {
-        this.pagination.currentPage += 1;
-        this.track.renderTrack(this.pagination.currentPage);
-        this.track.subtitle.textContent = this.pagination.setPage();
-      }
-    };
-    this.pagination.nextBtn.addEventListener('click', paginationNextHandler);
-    const paginationPrevHandler = (): void => {
-      if (this.pagination.currentPage > 1) {
-        this.pagination.currentPage -= 1;
-        this.track.renderTrack(this.pagination.currentPage);
-        this.track.subtitle.textContent = this.pagination.setPage();
-      }
-    };
-    this.pagination.prevBtn.addEventListener('click', paginationPrevHandler);
   }
 }
