+++ async-race/src/components/garage/track-component/types/track-types.ts
@@ -1,9 +1,11 @@
 import type { IParams } from '../../../../types/types';
+import type { IPagination } from '../../../pagination/types/pagination-types';
 
 export interface ITrack {
   title: HTMLElement
   subtitle: HTMLElement
   trackList: HTMLElement
+  pagination: IPagination
 }
 
 export interface ITrackParams {
